+++ src/index.php
@@ -807,6 +807,34 @@
                             onclick="tmp=parseInt($('#udpages').val());$('#udpages').val(tmp+1);loadUserDelivery();">></button>
                     </div>
                 </div>
+                <br>
+                <div class="py-8 px-6 pt-4 bg-white shadow rounded" id="userSessions" style="display:none">
+                    <div class="flex px-6 pb-4 border-b">
+                        <h2><b><?php echo $st->sessions; ?></b></h2>
+                        &nbsp;&nbsp;&nbsp;
+                        <button type="button" style="display:inline;padding:5px" id="revokeAllBtn"
+              class="w-full md:w-auto px-6 py-3 font-medium text-white bg-indigo-500 hover:bg-indigo-600 rounded transition duration-200"
+              onclick="revokeAllToken()">Revoke All</button>
+                    </div>
+                    <div class="p-4 overflow-x-auto" style="display: block;">
+                        <table class="table-auto w-full">
+                            <thead id="sessionTableHead">
+                                <tr class="text-xs text-gray-500 text-left">
+                                    <th class="py-5 px-6 pb-3 font-medium">IP</th>
+                                    <th class="py-5 px-6 pb-3 font-medium"><?php echo $st->login_time; ?></th>
+                                    <th class="py-5 px-6 pb-3 font-medium"><?php echo $st->expire_time; ?></th>
+                                </tr>
+                            </thead>
+                            <tbody id="sessiontable">
+                                <tr class="text-sm">
+                                    <td class="py-5 px-6 font-medium"><?php echo $st->no_data; ?></td>
+                                    <td class="py-5 px-6 font-medium"></td>
+                                    <td class="py-5 px-6 font-medium"></td>
+                                </tr>
+                            </tbody>
+                        </table>
+                    </div>
+                </div>
             </div>
             <div class="w-full md:w-1/3 px-4 mb-4 md:mb-0">
                 <div class="py-8 px-6 pt-4 bg-white shadow rounded" id="Account">
@@ -911,47 +939,9 @@
         <div class="px-4 mx-auto">
             <div class="flex flex-wrap -mx-4 -mb-4 md:mb-0 py-8 px-6">
                 <div class="md:w-2/3 px-4 mb-4 md:mb-0" id="hometableftcontainer">
-                    <div class="flex flex-wrap -mx-4 -mb-4 md:mb-0 px-6" id="HomeTabLeft">
-                        <div class="px-6" style="width:100%">
-                            <div class="p-6 border-l-4 border-2 border-indigo-500 rounded-r-lg bg-gray-50">
-                                <div class="flex">
-                                    <div class="w-auto">
-                                        <span class="w-auto inline-block mr-2">
-                                            <svg width="20" height="20" viewbox="0 0 20 20" fill="none"
-                                                xmlns="http://www.w3.org/2000/svg">
-                                                <path
-                                                    d="M10 0C4.5 0 0 4.5 0 10C0 15.5 4.5 20 10 20C15.5 20 20 15.5 20 10C20 4.5 15.5 0 10 0ZM11 14C11 14.6 10.6 15 10 15C9.4 15 9 14.6 9 14V10C9 9.4 9.4 9 10 9C10.6 9 11 9.4 11 10V14ZM10 7C9.4 7 9 6.6 9 6C9 5.4 9.4 5 10 5C10.6 5 11 5.4 11 6C11 6.6 10.6 7 10 7Z"
-                                                    fill="#382CDD"></path>
-                                            </svg>
-                                        </span>
-                                    </div>
-                                    <div class="w-full">
-                                        <div class="flex">
-                                            <h3 class="text-indigo-800 font-medium">Looking for staff!</h3>
-                                        </div>
-                                        <div class="pr-6">
-                                            <p class="text-sm text-indigo-700">
-                                                <span style='font-size:10px;color:grey'>Global announcement by
-                                                    CharlesWithC.</span><br>
-                                                <span style='font-size:10px;color:grey'>This is not a in-VTC
-                                                    announcement and this doesn't refer to looking for VTC
-                                                    staff.</span><br>
-                                                In order to make the <a href="https://drivershub.charlws.com"
-                                                    target="_blank">Drivers Hub Project</a> grow faster and bigger,
-                                                we're opening <b>staff positions</b>.<br>
-                                                <b>Community Helper</b> and <b>Marketing Team (PAID)</b> are
-                                                available<br>
-                                                <a style="font-size:16px;" href="https://go.charlws.com/hubstaff"
-                                                    target="_blank">Click to Apply</a>
-                                            </p>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <br>
-                    <div style="width:100%;">
+                    <!-- <div class="flex flex-wrap -mx-4 -mb-4 md:mb-0 px-6" id="HomeTabLeft">
+                    </div> -->
+                    <div id="statsTimeRange" style="width:100%;display:none">
                         <div style="margin-left:20px;margin-bottom:-10px">
                             <input id="stats_start"
                                 class="block w-full px-4 py-3 mb-2 text-sm placeholder-gray-500 bg-white border rounded"
@@ -975,7 +965,7 @@
                                                 d="M6 2a.5.5 0 0 1 .47.33L10 12.036l1.53-4.208A.5.5 0 0 1 12 7.5h3.5a.5.5 0 0 1 0 1h-3.15l-1.88 5.17a.5.5 0 0 1-.94 0L6 3.964 4.47 8.171A.5.5 0 0 1 4 8.5H.5a.5.5 0 0 1 0-1h3.15l1.88-5.17A.5.5 0 0 1 6 2Z"
                                                 fill="#382CDD"></path>
                                         </svg> </span>
-                                    <h3 class="text-sm text-gray-600"><?php echo $st->live; ?></h3><span
+                                    <h3 class="text-sm text-gray-600"><?php echo $st->live; ?></h3><span onclick="$('#statsTimeRange').fadeIn();"
                                         class="inline-block ml-auto px-2 py-1 text-xs text-gray-500 rounded-full"><?php echo $st->now; ?></span>
                                 </div>
                                 <h2 class="mb-2 text-3xl font-bold" id="livedriver">-</h2>
@@ -1004,7 +994,7 @@
                                             <line x1="10" y1="12" x2="3.25" y2="10" />
                                             <line x1="14" y1="12" x2="20.75" y2="10" />
                                         </svg> </span>
-                                    <h3 class="text-sm text-gray-600"><?php echo $st->drivers; ?></h3><span
+                                    <h3 class="text-sm text-gray-600"><?php echo $st->drivers; ?></h3><span onclick="$('#statsTimeRange').fadeIn();"
                                         class="inline-block ml-auto px-2 py-1 text-xs text-gray-500 rounded-full"><?php echo $st->all; ?></span>
                                 </div>
                                 <h2 class="mb-2 text-3xl font-bold" id="alldriver">-</h2>
@@ -1026,7 +1016,7 @@
                                                 d="M0 3.5A1.5 1.5 0 0 1 1.5 2h9A1.5 1.5 0 0 1 12 3.5V5h1.02a1.5 1.5 0 0 1 1.17.563l1.481 1.85a1.5 1.5 0 0 1 .329.938V10.5a1.5 1.5 0 0 1-1.5 1.5H14a2 2 0 1 1-4 0H5a2 2 0 1 1-3.998-.085A1.5 1.5 0 0 1 0 10.5v-7zm1.294 7.456A1.999 1.999 0 0 1 4.732 11h5.536a2.01 2.01 0 0 1 .732-.732V3.5a.5.5 0 0 0-.5-.5h-9a.5.5 0 0 0-.5.5v7a.5.5 0 0 0 .294.456zM12 10a2 2 0 0 1 1.732 1h.768a.5.5 0 0 0 .5-.5V8.35a.5.5 0 0 0-.11-.312l-1.48-1.85A.5.5 0 0 0 13.02 6H12v4zm-9 1a1 1 0 1 0 0 2 1 1 0 0 0 0-2zm9 0a1 1 0 1 0 0 2 1 1 0 0 0 0-2z"
                                                 fill="#382CDD"></path>
                                         </svg></span>
-                                    <h3 class="text-sm text-gray-600"><?php echo $st->delivered; ?></h3><span
+                                    <h3 class="text-sm text-gray-600"><?php echo $st->delivered; ?></h3><span onclick="$('#statsTimeRange').fadeIn();"
                                         class="inline-block ml-auto px-2 py-1 text-xs text-gray-500 rounded-full"><?php echo $st->all; ?></span>
                                 </div>
                                 <h2 class="mb-2 text-3xl font-bold" id="alljob">-</h2>
@@ -1050,7 +1040,7 @@
                                             <circle cx="12" cy="10" r="3"></circle>
                                         </svg>
                                     </span>
-                                    <h3 class="text-sm text-gray-600"><?php echo $st->distance; ?></h3><span
+                                    <h3 class="text-sm text-gray-600"><?php echo $st->distance; ?></h3><span onclick="$('#statsTimeRange').fadeIn();"
                                         class="inline-block ml-auto px-2 py-1 text-xs text-gray-500 rounded-full"><?php echo $st->all; ?></span>
                                 </div>
                                 <h2 class="mb-2 text-3xl font-bold" id="alldistance">-</h2>
@@ -1075,7 +1065,7 @@
                                             <path
                                                 d="M8 13.5a5.5 5.5 0 1 1 0-11 5.5 5.5 0 0 1 0 11zm0 .5A6 6 0 1 0 8 2a6 6 0 0 0 0 12z" />
                                         </svg> </span>
-                                    <h3 class="text-sm text-gray-600"><?php echo $st->profit; ?></h3><span
+                                    <h3 class="text-sm text-gray-600"><?php echo $st->profit; ?></h3><span onclick="$('#statsTimeRange').fadeIn();"
                                         class="inline-block ml-auto px-2 py-1 text-xs text-gray-500 rounded-full"><?php echo $st->all; ?></span>
                                 </div>
                                 <h2 class="mb-2 text-3xl font-bold" id="allprofit">-</h2>
@@ -1098,7 +1088,7 @@
                                                 d="M13,6L13,6v5.5c0,0.2761-0.2239,0.5-0.5,0.5S12,11.7761,12,11.5v-2C12,8.6716,11.3284,8,10.5,8H9V2c0-0.5523-0.4477-1-1-1H2 C1.4477,1,1,1.4477,1,2v11c0,0.5523,0.4477,1,1,1h6c0.5523,0,1-0.4477,1-1V9h1.5C10.7761,9,11,9.2239,11,9.5v2 c0,0.8284,0.6716,1.5,1.5,1.5s1.5-0.6716,1.5-1.5V5c0-0.5523-0.4477-1-1-1l0,0V2.49C12.9946,2.2178,12.7723,1.9999,12.5,2 c-0.2816,0.0047-0.5062,0.2367-0.5015,0.5184C11.9987,2.5289,11.9992,2.5395,12,2.55V5C12,5.5523,12.4477,6,13,6s1-0.4477,1-1 s-0.4477-1-1-1 M8,6.5C8,6.7761,7.7761,7,7.5,7h-5C2.2239,7,2,6.7761,2,6.5v-3C2,3.2239,2.2239,3,2.5,3h5C7.7761,3,8,3.2239,8,3.5 V6.5z"
                                                 fill="#382CDD"></path>
                                         </svg> </span>
-                                    <h3 class="text-sm text-gray-600"><?php echo $st->fuel; ?></h3><span
+                                    <h3 class="text-sm text-gray-600"><?php echo $st->fuel; ?></h3><span onclick="$('#statsTimeRange').fadeIn();"
                                         class="inline-block ml-auto px-2 py-1 text-xs text-gray-500 rounded-full"><?php echo $st->all; ?></span>
                                 </div>
                                 <h2 class="mb-2 text-3xl font-bold" id="allfuel">-</h2>
@@ -3128,7 +3118,8 @@
 
                         <div class="mb-6">
                             <label class="block text-sm font-medium mb-2" for="">Types (Separate types with
-                                line-break / For each type, separate id, name, discord_role_id, staff_role_id (separate with '|'), message, webhook, note with ',' /
+                                line-break / For each type, separate id, name, discord_role_id, staff_role_id (separate
+                                with '|'), message, webhook, note with ',' /
                                 ID 1~4 are reserved and name change will not take effect)</label>
                             <textarea id="config_application_types_txt" style="width:100%;height:100px"
                                 class="block w-full px-4 py-3 mb-2 text-sm placeholder-gray-500 bg-white border rounded"
@@ -3228,7 +3219,7 @@
                 API: <span id="apiversion">v?.?.?</span> <a href="https://drivershub.charlws.com/changelog"
                     target="_blank">Changelog</a>
                 &nbsp;|&nbsp;
-                Web: v1.3.3 <a href="/changelog" target="_blank">Changelog</a>
+                Web: v1.3.4 <a href="/changelog" target="_blank">Changelog</a>
                 <br>
                 Map: <a href="https://map.charlws.com" target="_blank">map.charlws.com</a>
                 &nbsp;|&nbsp;